Revert "89 topbar component"
