+++ backend/src/app.ts
@@ -120,12 +120,6 @@ export default class App {
       return res.json(album);
     });
 
-    this.server.get('/api/spotify/volume', async (req: Request, res: Response) => {
-      const volume: any = req.query?.volume ?? 100;
-      const result = await this.spotifyService.setVolume(volume);
-      return res.json(result);
-    });
-
     // DB
     // TODO
