+++ backend/src/connectors/spotifyService.ts
@@ -130,10 +130,5 @@ export default class SpotifyService {
     return album.body;
   }
 
-  setVolume = async (volume: number) => {
-    const result = await this.spotifyApi.setVolume(volume);
-    return result;
-  }
-
   getAccessToken = async () => this.spotifyApi.getAccessToken();
 }
