+++ frontend/package.json
@@ -12,7 +12,6 @@
     "@types/react-dom": "^17.0.10",
     "exenv": "^1.2.2",
     "node-sass": "^6.0.1",
-    "rc-slider": "^9.7.5",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-grid-layout": "^1.3.0",
