+++ frontend/src/App.scss
@@ -1,5 +1,11 @@
 @import "boilerplate";
 
+.Searchbar {
+  position: fixed;
+  top: 0;
+  z-index: 1;
+}
+
 .AuthorizeButton {
   position: fixed;
   top: 0;
