+++ frontend/src/App.tsx
@@ -3,13 +3,23 @@ import React, { useEffect, useState } from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import Albums from "./components/Albums/Albums";
 import Playlists from "./components/Playlists/Playlists";
+import Searchbar from "./components/Searchbar/Searchbar";
 import SavedTracks from "./components/SavedTracks/SavedTracks";
 import Player from "./components/Player/Player";
+import { API_URL } from "./utils/constants";
 import Sidebar from "./components/Sidebar/Sidebar";
 import Dashboard from "./components/Dashboard/Dashboard";
 import PlaylistPage from "./pages/PlaylistPage/PlaylistPage";
 import AlbumPage from "./pages/AlbumPage/AlbumPage";
-import Topbar from "./components/Topbar/Topbar";
+
+function authorize() {
+  fetch(`${API_URL}api/spotify/get-auth-url`)
+    .then((res) => res.text())
+    .then((url) => {
+      console.log(url);
+      window.location.href = url;
+    });
+}
 
 function App() {
   const [isAuthorized, setIsAuthorized] = useState(false);
@@ -78,7 +88,8 @@ function App() {
             </Route>
           </Switch>
         </div>
-      </Router>
+      </div>
+    </Router>
   );
 }
