+++ frontend/src/components/Searchbar/Searchbar.scss
@@ -1,39 +1,11 @@
-@import './src/boilerplate.scss';
-
 .Searchbar {
 
-  color: $white;
 
-  .SearchbarInput {
-    width: 80%;
-    background: $color-accent-dark;
-    padding: 9px 20px 9px 50px;
-    border: none;
-    border-radius: 50px;
-  }
+  .SearchbarInput {}
 
-  .SearchbarInput::placeholder {
-    color: $gray-light;
-  }
+  .SearchbarButton {}
 
-  .search-icon {
-    position: absolute;
-    color: $gray-light;
-    padding: 6px 16px;
-  }
+  .SearchbarResults {}
 
-  .SearchbarResults {
-    position: absolute;
-    padding: 10px 20px;
-    list-style-type: none;
-    background: $color-accent-dark;
-  }
-
-  .SearchbarResultItem {
-    padding: 5px 0;
-
-    img {
-      margin-right: 10px;
-    }
-  }
+  .SearchbarResultItem {}
 }
