--- frontend/src/components/Topbar/Topbar.scss
@@ -1,73 +1,0 @@
-@import './src/boilerplate.scss';
-
-.top-bar {
-  position: sticky;
-  top: 0;
-  z-index: 3;
-  background: $color-background;
-  padding: 32px;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-
-  .top-bar-item {
-    display: flex;
-  }
-
-  .search {
-    align-items: center;
-    justify-content: flex-start;
-    flex-grow: 1;
-    flex-basis: 400px;
-
-    :first-child {
-      flex-grow: 1;
-    }
-  }
-
-  .volume {
-    flex-grow: 1;
-    flex-basis: 300px;
-    color: $gray-light;
-    align-items: center;
-    justify-content: flex-end;
-
-    .settings-button {
-      padding: 4px 7px;
-      background: transparent;
-      border: none;
-    }
-
-    .volume-slider {
-      display: inline-block;
-      width: 50%;
-      margin-left: 10px;
-
-      .rc-slider-track, .rc-slider-handle {
-        background-color: $gray-light;
-        border-color: $gray-light;
-      }
-    }
-  }
-
-  .settings {
-    flex-grow: 1;
-    flex-basis: 200px;
-    align-items: center;
-    justify-content: flex-end;
-
-    .settings-button {
-      border-radius: 10px;
-      padding: 4px 7px;
-      background: $color-accent-dark;
-      border: none;
-      float: right;
-      margin-left: 20px;
-
-      .material-icons {
-        color: $gray-light;
-      }
-    }
-  }
-
-}
