+++ frontend/src/index.scss
@@ -27,7 +27,7 @@
     height: 100vh;
     flex: 1 1 auto;
     overflow: hidden auto;
-    padding: 32px $padding-horizontal;
+    padding: 64px $padding-horizontal;
     z-index: 0;
   }
